requirements.txt
langchain-community
langchain-huggingface
sentence-transformers
chromadb
numpy
python-dotenv
openai
pdfplumber
pdf2image
Pillow
pytesseract
python-docx
python-pptx
beautifulsoup4
requests
trafilatura
fastapi
uvicorn[standard]
python-multipart
tavily-python
duckduckgo-search
pyyaml
jinja2
jieba
